Migrate from LTS jenkins image to weekly release
I've noticed that this repository has started to get backed up with
plugins that are unable to be updated due to the Jenkins version of the
LTS image being too "old". This causes Renovate to be unable to update
other software dependencies.

The solution to work with is to forsake the LTS base image of Jenkins
and to use the weekly release but have Renovate only update the image
every 2 weeks. This technically only applies every second weekly release
but I feel this gives me sufficient air room when applying updates.

Another solution was to only update plugins every 12-16 weeks on a
single day. The idea is, this would be after a new baseline release of
the LTS Jenkins image but this would be error prone (e.g. a lot could go
wrong on a single day) and too slow for my liking.