Deprecate the current automatic versioning
The solution made to do automatic versioning for this repository (and
the original jenkins-base repository) is overkill for just wanting to be
consistent when it comes to small amount of tagging done. Hence, at this
point I don't find it worth maintaining.

From this commit onward, tagging of the repository will be done
judiciously. Any patches done through Renovate will not cause any new
version change as ideally this should be only applicable when the
"casc.yaml" changes.

That said, I still want images to be built weekly and for me to still
have the ability to create an image via push.