Migrate this repository's name to jenkins-homelab
This was done because I've wanted the name of the repository to reflect
what the Jenkins instance is for.

The description within the pyproject.toml has changed because typically,
I have this field match what is in the 'README.md'.

The README.md description has changed because it is uncertain whether
this instance will do much CI/CD for my source code repositories. Hence,
Hence, I've decided on a description that is more generic.